# docs/event-service-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
...
    "title" : "Event Service API",
    "version" : "1.0"
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/api/events" : {
      "get" : {
        "description" : "Everyone sees PUBLISHED events. Organizers also see their own non-published events. ADMIN and ADMINISTRATIVE_STAFF see all. All filters are optional and combine with AND; results are ordered by start time. Without page/size the whole filtered list is returned. X-Total-Count always holds the number of matching events.",
        "operationId" : "listEvents",
        "parameters" : [ {
          "description" : "Only events with this status",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "DRAFT", "PUBLISHED", "CANCELLED", "COMPLETED" ]
          }
        }, {
          "description" : "Only events starting at or after this time",
          "example" : "2026-10-01T00:00:00",
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "description" : "Only events starting at or before this time",
          "example" : "2026-10-31T23:59:59",
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "description" : "Only events that have not started yet",
          "in" : "query",
          "name" : "upcoming",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "description" : "Only events the caller organizes",
          "in" : "query",
          "name" : "mine",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "description" : "Page number, from 0",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Page size, 1 to 100 (default 20 when page is given)",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EventResponse"
                  }
                }
              }
            },
            "description" : "Visible events",
            "headers" : {
              "X-Total-Count" : {
                "description" : "Number of events matching the filters (all pages)",
                "schema" : {
                  "type" : "integer",
                  "format" : "int64"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_PARAMETER (unknown status or badly formatted date) or VALIDATION_ERROR (to before from, or page/size out of range)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "List events visible to the caller",
        "tags" : [ "Events" ]
      },
      "post" : {
        "description" : "Roles: EVENT_ORGANIZER, ACADEMIC_STAFF, ADMIN. The caller becomes the event's organizer. eligibilityRule is {\"all\": true} or Group 5 criteria such as {\"roles\": [\"STUDENT\"], \"departmentId\": \"dep-cs\"}.",
        "operationId" : "createEvent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            },
            "description" : "Event created in DRAFT status"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "VALIDATION_ERROR: invalid or missing fields, scheduleEnd not after scheduleStart, or registrationCloseAt after scheduleStart. INVALID_ELIGIBILITY_RULE: eligibilityRule is not in the supported format. MALFORMED_REQUEST: body is not valid JSON"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED: missing, invalid or expired token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller is not EVENT_ORGANIZER, ACADEMIC_STAFF or ADMIN"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Create a draft event",
        "tags" : [ "Events" ]
      }
    },
...
    "/api/events/{id}" : {
      "get" : {
        "operationId" : "getEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            },
            "description" : "Event detail"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_PARAMETER: id is not a UUID"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_VISIBLE: event is not published and caller is neither its organizer nor ADMIN / ADMINISTRATIVE_STAFF"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Get event detail",
        "tags" : [ "Events" ]
      },
      "patch" : {
        "description" : "Only provided fields change. Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN.",
        "operationId" : "updateEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            },
            "description" : "Updated event"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "VALIDATION_ERROR, INVALID_SCHEDULE (resulting schedule is inconsistent), INVALID_ELIGIBILITY_RULE, MALFORMED_REQUEST (body is not valid JSON) or INVALID_PARAMETER (id is not a UUID)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller does not own the event"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Partially update an event",
        "tags" : [ "Events" ]
      }
    },
    "/api/events/{id}/cancel" : {
      "patch" : {
        "description" : "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN.",
        "operationId" : "cancelEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            },
            "description" : "Event is now CANCELLED"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_PARAMETER (id is not a UUID) or INVALID_STATE (event is already cancelled or completed)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller does not own the event"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Cancel an event",
        "tags" : [ "Events" ]
      }
    },
    "/api/events/{id}/complete" : {
      "patch" : {
        "description" : "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN. For events that finish early: published events are also completed automatically once their end time passes. Feedback is only accepted for COMPLETED events.",
        "operationId" : "completeEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            },
            "description" : "Event is now COMPLETED"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_PARAMETER (id is not a UUID), INVALID_STATE (only PUBLISHED events can be completed) or EVENT_NOT_STARTED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller does not own the event"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Mark a published event as completed",
        "tags" : [ "Events" ]
      }
    },
    "/api/events/{id}/publish" : {
      "patch" : {
        "description" : "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN. Physical-venue events are checked with Group 6 first; online events skip the check.",
        "operationId" : "publishEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventResponse"
                }
              }
            },
            "description" : "Event is now PUBLISHED"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_PARAMETER (id is not a UUID), INVALID_STATE (only DRAFT events can be published) or VENUE_NOT_FOUND: Group 6 does not know the venue code"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller does not own the event"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "VENUE_NOT_AVAILABLE: Group 6 says the venue cannot be used"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "GROUP6_UNAVAILABLE: venue service unreachable, event stays DRAFT"
          }
        },
        "summary" : "Publish a draft event",
        "tags" : [ "Events" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "EventResponse" : {
        "type" : "object",
        "properties" : {
          "capacity" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 30
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string",
            "example" : "Hands-on workshop for Computing students"
          },
          "eligibilityRule" : {
            "type" : "string",
            "description" : "JSON rule as a string, e.g. {\"all\": true} or {\"roles\": [\"STUDENT\"], \"departmentId\": \"dep-cs\"}",
            "example" : {
              "roles" : [ "STUDENT" ],
              "departmentId" : "dep-cs"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "a0000000-0000-0000-0000-000000000001"
          },
          "online" : {
            "type" : "boolean",
            "example" : false
          },
          "organizerId" : {
            "type" : "string",
            "description" : "Group 5 user id of the organizer",
            "example" : "usr-organizer-001"
          },
          "registrationCloseAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-10-04T18:00:00"
          },
          "registrationOpenAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-09-25T08:00:00"
          },
          "scheduleEnd" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-10-05T13:00:00"
          },
          "scheduleStart" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-10-05T10:00:00"
          },
          "status" : {
            "type" : "string",
            "description" : "DRAFT -> PUBLISHED -> COMPLETED, or CANCELLED from DRAFT/PUBLISHED",
            "enum" : [ "DRAFT", "PUBLISHED", "CANCELLED", "COMPLETED" ]
          },
          "title" : {
            "type" : "string",
            "example" : "Innovation Week Workshop"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "venue" : {
            "type" : "string",
            "description" : "Group 6 resource code; null for online events",
            "example" : "LAB-101"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}