# src/main/java/com/group8/eventservice/controller/EventController.java
package com.group8.eventservice.controller;

import java.time.LocalDateTime;
import java.util.List;
import java.util.UUID;

import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import com.group8.eventservice.dto.request.CreateEventRequest;
import com.group8.eventservice.dto.request.EventFilter;
import com.group8.eventservice.dto.request.UpdateEventRequest;
import com.group8.eventservice.dto.response.EventResponse;
import com.group8.eventservice.entity.EventStatus;
import com.group8.eventservice.exception.ApiErrorResponse;
import com.group8.eventservice.service.EventService;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.headers.Header;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/api/events")
@RequiredArgsConstructor
@Tag(name = "Events", description = "Create, edit, publish, cancel and view events")
public class EventController {

    private static final String ERR = "application/json";
    private static final String TOTAL_COUNT = "X-Total-Count";
    private static final String MANAGE_ROLES = "hasAnyRole('EVENT_ORGANIZER', 'ACADEMIC_STAFF', 'ADMIN')";

    private final EventService eventService;

    @PostMapping
    @PreAuthorize(MANAGE_ROLES)
    @Operation(summary = "Create a draft event", description = "Roles: EVENT_ORGANIZER, ACADEMIC_STAFF, ADMIN. The caller becomes the event's organizer. "
                    + "eligibilityRule is {\"all\": true} or Group 5 criteria such as {\"roles\": [\"STUDENT\"], \"departmentId\": \"dep-cs\"}.")
    @ApiResponse(responseCode = "201", description = "Event created in DRAFT status")
    @ApiResponse(responseCode = "400", description = "VALIDATION_ERROR: invalid or missing fields, scheduleEnd not after scheduleStart, or registrationCloseAt after scheduleStart. "
            + "INVALID_ELIGIBILITY_RULE: eligibilityRule is not in the supported format. MALFORMED_REQUEST: body is not valid JSON",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED: missing, invalid or expired token",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "FORBIDDEN: caller is not EVENT_ORGANIZER, ACADEMIC_STAFF or ADMIN",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public ResponseEntity<EventResponse> createEvent(@Valid @RequestBody CreateEventRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(eventService.createEvent(request));
    }

    @GetMapping
    @Operation(summary = "List events visible to the caller",
            description = "Everyone sees PUBLISHED events. Organizers also see their own non-published events. "
                    + "ADMIN and ADMINISTRATIVE_STAFF see all. All filters are optional and combine with AND; results are "
                    + "ordered by start time. Without page/size the whole filtered list is returned. "
                    + "X-Total-Count always holds the number of matching events.")
    @ApiResponse(responseCode = "200", description = "Visible events",
            headers = @Header(name = TOTAL_COUNT, description = "Number of events matching the filters (all pages)",
                    schema = @Schema(implementation = Long.class, example = "57")))
    @ApiResponse(responseCode = "400", description = "INVALID_PARAMETER (unknown status or badly formatted date) or "
            + "VALIDATION_ERROR (to before from, or page/size out of range)",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public ResponseEntity<List<EventResponse>> listEvents(
            @Parameter(description = "Only events with this status") @RequestParam(required = false) EventStatus status,
            @Parameter(description = "Only events starting at or after this time", example = "2026-10-01T00:00:00")
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime from,
            @Parameter(description = "Only events starting at or before this time", example = "2026-10-31T23:59:59")
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime to,
            @Parameter(description = "Only events that have not started yet") @RequestParam(defaultValue = "false") boolean upcoming,
            @Parameter(description = "Only events the caller organizes") @RequestParam(defaultValue = "false") boolean mine,
            @Parameter(description = "Page number, from 0") @RequestParam(required = false) Integer page,
            @Parameter(description = "Page size, 1 to 100 (default 20 when page is given)") @RequestParam(required = false) Integer size) {
        EventService.EventPage result = eventService.listVisibleEvents(
                new EventFilter(status, from, to, upcoming, mine, page, size));
        return ResponseEntity.ok()
                .header(TOTAL_COUNT, String.valueOf(result.total()))
                .body(result.events());
    }

    @GetMapping("/{id}")
    @Operation(summary = "Get event detail")
    @ApiResponse(responseCode = "200", description = "Event detail")
    @ApiResponse(responseCode = "400", description = "INVALID_PARAMETER: id is not a UUID",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "NOT_VISIBLE: event is not published and caller is neither its organizer nor ADMIN / ADMINISTRATIVE_STAFF",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "NOT_FOUND",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public EventResponse getEvent(@PathVariable UUID id) {
        return eventService.getEventDetail(id);
    }

    @PatchMapping("/{id}")
    @PreAuthorize(MANAGE_ROLES)
    @Operation(summary = "Partially update an event", description = "Only provided fields change. Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN.")
    @ApiResponse(responseCode = "200", description = "Updated event")
    @ApiResponse(responseCode = "400", description = "VALIDATION_ERROR, INVALID_SCHEDULE (resulting schedule is inconsistent), INVALID_ELIGIBILITY_RULE, "
            + "MALFORMED_REQUEST (body is not valid JSON) or INVALID_PARAMETER (id is not a UUID)",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "FORBIDDEN: caller does not own the event",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "NOT_FOUND",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public EventResponse updateEvent(@PathVariable UUID id, @Valid @RequestBody UpdateEventRequest request) {
        return eventService.updateEvent(id, request);
    }

    @PatchMapping("/{id}/publish")
    @PreAuthorize(MANAGE_ROLES)
    @Operation(summary = "Publish a draft event",
            description = "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN. Physical-venue events are checked with Group 6 first; online events skip the check.")
    @ApiResponse(responseCode = "400", description = "INVALID_PARAMETER (id is not a UUID), INVALID_STATE (only DRAFT events can be published) or VENUE_NOT_FOUND: Group 6 does not know the venue code",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "409", description = "VENUE_NOT_AVAILABLE: Group 6 says the venue cannot be used",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "503", description = "GROUP6_UNAVAILABLE: venue service unreachable, event stays DRAFT",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "200", description = "Event is now PUBLISHED")
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "FORBIDDEN: caller does not own the event",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "NOT_FOUND",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public EventResponse publishEvent(@PathVariable UUID id) {
        return eventService.publishEvent(id);
    }

    @PatchMapping("/{id}/cancel")
    @PreAuthorize(MANAGE_ROLES)
    @Operation(summary = "Cancel an event", description = "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN.")
    @ApiResponse(responseCode = "200", description = "Event is now CANCELLED")
    @ApiResponse(responseCode = "400", description = "INVALID_PARAMETER (id is not a UUID) or INVALID_STATE (event is already cancelled or completed)",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "FORBIDDEN: caller does not own the event",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "NOT_FOUND",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public EventResponse cancelEvent(@PathVariable UUID id) {
        return eventService.cancelEvent(id);
    }

    @PatchMapping("/{id}/complete")
    @PreAuthorize(MANAGE_ROLES)
    @Operation(summary = "Mark a published event as completed",
            description = "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN. For events that finish early: "
                    + "published events are also completed automatically once their end time passes. Feedback is only accepted for COMPLETED events.")
    @ApiResponse(responseCode = "200", description = "Event is now COMPLETED")
    @ApiResponse(responseCode = "400", description = "INVALID_PARAMETER (id is not a UUID), INVALID_STATE (only PUBLISHED events can be completed) or EVENT_NOT_STARTED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "UNAUTHORIZED",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "FORBIDDEN: caller does not own the event",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "NOT_FOUND",
            content = @Content(mediaType = ERR, schema = @Schema(implementation = ApiErrorResponse.class)))
    public EventResponse completeEvent(@PathVariable UUID id) {
        return eventService.completeEvent(id);
    }
}